+++ .github/workflows/auto_update.yml
@@ -21,7 +21,7 @@ jobs:
 
       - name: Setup git config
         run: |
-          git config user.name "GitHub Actions Bot"
+          git config user.name "CyberOwl Bot"
           git config user.email "<>"
 
       - name: Commit
