+++ src/cyberowl/pipelines.py
@@ -7,9 +7,11 @@ from mdtemplate import Template
 class AlertPipeline:
     """
     AlertPipeline class
+    Args:
+        items : list of items.
     """
 
-    result: list = []
+    items: list = []
 
     def remove_special_characters(self, text):
         """
@@ -23,13 +25,13 @@ class AlertPipeline:
             .replace("|", "")
         )
 
-    def open_spider(self, spider):
+    def open_spider(self, *args, **kwargs):
         """
         Open spider
         """
-        self.result = []
+        self.items = []
 
-    def process_item(self, item, spider):
+    def process_item(self, item, *args, **kwargs):
         """
         Process item.
         """
@@ -38,13 +40,13 @@ class AlertPipeline:
         item["date"] = self.remove_special_characters(item["date"])
         item["description"] = self.remove_special_characters(item["description"])
 
-        self.result.append(item)
+        self.items.append(item)
 
     def close_spider(self, spider):
         """
         Close spider
         """
-        to_write = Template(spider.name, self.result)
+        to_write = Template(spider.name, self.items)
         with open("README.md", "a", encoding="utf-8") as file:
             file.write(to_write.fill_table())
             file.close()
