+++ src/cyberowl/spiders/cert_fr_spider.py
@@ -1,5 +1,4 @@
 import scrapy
-
 from items import AlertItem
