--- src/items.py
@@ -1,17 +1,0 @@
-"""
-    Scrapy Item classes
-"""
-
-import scrapy
-
-
-class AlertItem(scrapy.Item):
-    """
-    AlertItem class.
-    """
-
-    source = scrapy.Field()
-    link = scrapy.Field()
-    title = scrapy.Field()
-    description = scrapy.Field()
-    date = scrapy.Field()
